Split geos package into more files
This is because it was starting to get unwieldy and unorganized. It's
now better separated into layers (entrypoints, generic operations, and
the handles).
