+++ geos/entrypoints_test.go
