+++ geos/errors.go
@@ -1,0 +1,10 @@
+package geos
+
+import "fmt"
+
+func wrap(err error, format string, args ...interface{}) error {
+	if err == nil {
+		return nil
+	}
+	return fmt.Errorf(format+": %v", append(args, err)...)
+}
